Create header for common, private iokit backend functions.
Currently, the iokit backend only allows one DeviceInterface per
IOKit class. But we want both Block and StorageVolume to use the
'IOMedia' class, so we have to allow multiple DeviceInterfaces
per one IOKit class.

The typeFromEntry() mapping of IOKit-class => DeviceInterface
doesn't make sense for a many-to-one relationship, so we'll need
to use typeToName() for both directions. Since both
iokitmanager.cpp and iokitdevice.cpp will use this function, it
should be available in a common, private header.

While we're at it, move q_toVariantMap() to iokitprivate,
also namespace it and its helpers appropriately.