Add icons for iokit devices.
We currently don't distinguish between different sorts of partitions,
it would be nice to display different icons for USB disks and such.

It would also be nice to display the actual OS X icons for the disks,
as accessible through QFileIconProvider. But such icons may be
custom, while Device::icon() returns only a QString for a themed
icon, so it can't be done.