Work around bug in IORegistryEntryFromPath, in iokit backend.
Apple's IORegistryEntryFromPath() fails to return a registry entry
in some cases, specifically involving CoreStorage volumes. This bug
has been reported to Apple as rdar://10104777 .

Although it's their bug, it makes Solid crash when listing
StorageVolume devices, and it's easy to work around the problem. If
IORegistryEntryFromPath() fails, we simply navigate the IORegistry
ourselves. First we find an ancestor of the desired entry that can
be found with IORegistryEntryFromPath(), and then we look through
its children until we find the desired path.